+++ changelog.md
@@ -1,1 +1,17 @@
-Initial entry
+# Change Log
+Changes to the lab will be detailed in this log file.
+
+#### Feb 02 2016
+- initial entry
+- fixed first link and added formatting to changelog file
+- fixed second link
+- fixed first image link (sleep)
+
+#### Feb 04 2016
+- fixed second image link (lockemup)
+- fixed third image link (show)
+
+####Feb 05 2015
+- fixed fourth image link (dunno)
+- fixed fifth image link (bingo)
+- fixed sixth image link (wisdom)